+++ eslint.config.js
@@ -160,9 +160,7 @@ module.exports = [
       'jsdoc/require-returns': ['off'],
       // Ensure there is a blank line between the body and any @tags,
       // as required by the tsdoc spec.
-      // TODO: Re-enable soon.
-      // 'jsdoc/tag-lines': ['error', 'any', {'startLines': 1}],
-      'jsdoc/tag-lines': 'off',
+      'jsdoc/tag-lines': ['error', 'any', {'startLines': 1}],
 
       // Already handled by tsc.
       'no-dupe-class-members': 'off',