+++ plugins/block-dynamic-connection/src/dynamic_text_join.ts
@@ -42,6 +42,7 @@ const DYNAMIC_TEXT_JOIN_MIXIN = {
 
   /**
    * Create XML to represent number of text inputs.
+   *
    * @returns XML storage element.
    */
   mutationToDom(this: DynamicTextJoinBlock): Element {
@@ -59,6 +60,7 @@ const DYNAMIC_TEXT_JOIN_MIXIN = {
 
   /**
    * Parse XML to restore the text inputs.
+   *
    * @param xmlElement XML storage element.
    */
   domToMutation(this: DynamicTextJoinBlock, xmlElement: Element): void {
@@ -71,6 +73,7 @@ const DYNAMIC_TEXT_JOIN_MIXIN = {
 
   /**
    * Parses XML based on the 'inputs' attribute (non-standard).
+   *
    * @param xmlElement XML storage element.
    */
   deserializeInputs(this: DynamicTextJoinBlock, xmlElement: Element): void {
@@ -85,6 +88,7 @@ const DYNAMIC_TEXT_JOIN_MIXIN = {
 
   /**
    * Parses XML based on the 'items' attribute (standard).
+   *
    * @param xmlElement XML storage element.
    */
   deserializeCounts(this: DynamicTextJoinBlock, xmlElement: Element): void {
@@ -100,6 +104,7 @@ const DYNAMIC_TEXT_JOIN_MIXIN = {
 
   /**
    * Returns the state of this block as a JSON serializable object.
+   *
    * @returns The state of this block, ie the item count.
    */
   saveExtraState: function (this: DynamicTextJoinBlock): {itemCount: number} {
@@ -117,6 +122,7 @@ const DYNAMIC_TEXT_JOIN_MIXIN = {
 
   /**
    * Applies the given state to this block.
+   *
    * @param state The state to apply to this block, ie the item count.
    */
   loadExtraState: function (
@@ -137,6 +143,7 @@ const DYNAMIC_TEXT_JOIN_MIXIN = {
 
   /**
    * Check whether a new input should be added and determine where it should go.
+   *
    * @param connection The connection that has a pending connection.
    * @returns The index before which to insert a new input, or null if no input
    *     should be added.
@@ -179,6 +186,7 @@ const DYNAMIC_TEXT_JOIN_MIXIN = {
   /**
    * Called by a monkey-patched version of InsertionMarkerManager when
    * a block is dragged over one of the connections on this block.
+   *
    * @param connection The connection on this block that has a pending
    *     connection.
    */
@@ -218,6 +226,7 @@ const DYNAMIC_TEXT_JOIN_MIXIN = {
    * Filters the given target connections so that empty connections are removed,
    * unless we need those to reach the minimum input count. Empty connections
    * are removed starting at the end of the array.
+   *
    * @param targetConns The list of connections associated with inputs.
    * @returns A filtered list of connections (or null/undefined) which should
    *     be attached to inputs.
@@ -240,6 +249,7 @@ const DYNAMIC_TEXT_JOIN_MIXIN = {
    * every entry in the array (if it does not already exist). If the entry is
    * a connection and not null/undefined the connection will be connected to
    * the input.
+   *
    * @param targetConns The connections defining the inputs to add.
    */
   addItemInputs(
@@ -260,6 +270,7 @@ const DYNAMIC_TEXT_JOIN_MIXIN = {
 
   /**
    * Adds the top input with the label to this block.
+   *
    * @returns The added input.
    */
   addFirstInput(this: DynamicTextJoinBlock): Blockly.Input {