+++ plugins/block-shareable-procedures/src/i_procedure_block.ts
@@ -17,6 +17,7 @@ export interface IProcedureBlock {
 
 /**
  * A type guard which checks if the given block is a procedure block.
+ *
  * @param block The block to check for procedure-y-ness.
  * @returns Whether this block is a procedure block or not.
  */