+++ plugins/dev-create/templates/typescript-block/template/test/index.ts
@@ -17,6 +17,7 @@ const allBlocks = ['block_template'];
 
 /**
  * Create a workspace.
+ *
  * @param blocklyDiv The blockly container div.
  * @param options The Blockly options.
  * @returns The created workspace.