+++ plugins/dev-tools/src/index.d.ts
@@ -64,6 +64,7 @@ declare namespace DevTools {
 
   /**
    * Create the Blockly playground.
+   *
    * @param container
    * @param createWorkspace
    * @param defaultOptions
@@ -82,6 +83,7 @@ declare namespace DevTools {
 
   /**
    * Use dat.GUI to add controls to adjust configuration of a Blockly workspace.
+   *
    * @param createWorkspace
    * @param defaultOptions
    * @returns The dat.GUI instance.
@@ -94,6 +96,7 @@ declare namespace DevTools {
   /**
    * Generates a number of field testing blocks for a specific field and returns
    * the toolbox xml string.
+   *
    * @param fieldName
    * @param options
    * @returns The toolbox xml string.