+++ plugins/field-angle/test/index.ts
@@ -67,6 +67,7 @@ const toolbox = generateFieldTestBlocks('field_angle', [
 
 /**
  * Create a workspace.
+ *
  * @param blocklyDiv The blockly container div.
  * @param options The Blockly options.
  * @returns The created workspace.