+++ plugins/field-multilineinput/test/index.ts
@@ -36,6 +36,7 @@ All mimsy were the borogoves,
 
 /**
  * Create a workspace.
+ *
  * @param blocklyDiv The blockly container div.
  * @param options The Blockly options.
  * @returns The created workspace.