+++ plugins/renderer-inline-row-separators/test/index.ts
@@ -91,6 +91,7 @@ const toolbox: Blockly.utils.toolbox.ToolboxDefinition = {
 
 /**
  * Create a workspace.
+ *
  * @param blocklyDiv The blockly container div.
  * @param options The Blockly options.
  * @returns The created workspace.