+++ plugins/shadow-block-converter/test/index.ts
@@ -113,6 +113,7 @@ const toolbox: Blockly.utils.toolbox.ToolboxDefinition = {
 
 /**
  * Create a workspace.
+ *
  * @param blocklyDiv The blockly container div.
  * @param options The Blockly options.
  * @returns The created workspace.