+++ plugins/toolbox-search/test/index.ts
@@ -14,6 +14,7 @@ import '../src/toolbox_search';
 
 /**
  * Create a workspace.
+ *
  * @param blocklyDiv The blockly container div.
  * @param options The Blockly options.
  * @returns The created workspace.