+++ plugins/workspace-search/src/workspace_search.ts
@@ -75,6 +75,7 @@ export class WorkspaceSearch implements Blockly.IPositionable {
 
   /**
    * Class for workspace search.
+   *
    * @param workspace The workspace the search bar sits in.
    */
   constructor(private workspace: Blockly.WorkspaceSvg) {}
@@ -188,6 +189,7 @@ export class WorkspaceSearch implements Blockly.IPositionable {
 
   /**
    * Helper method for adding an event.
+   *
    * @param node Node upon which to listen.
    * @param name Event name to listen to (e.g. 'mousedown').
    * @param thisObject The value of 'this' in the function.
@@ -211,6 +213,7 @@ export class WorkspaceSearch implements Blockly.IPositionable {
   /**
    * Add a button to the action div. This must be called after the init function
    * has been called.
+   *
    * @param btn The button to add the event listener to.
    * @param onClickFn The function to call when the user clicks on
    *     or hits enter on the button.
@@ -222,6 +225,7 @@ export class WorkspaceSearch implements Blockly.IPositionable {
 
   /**
    * Creates the text input for the search bar.
+   *
    * @returns A text input for the search bar.
    */
   protected createTextInput(): HTMLInputElement {
@@ -233,6 +237,7 @@ export class WorkspaceSearch implements Blockly.IPositionable {
 
   /**
    * Creates the button used to get the next block in the list.
+   *
    * @returns The next button.
    */
   protected createNextBtn(): HTMLButtonElement {
@@ -241,6 +246,7 @@ export class WorkspaceSearch implements Blockly.IPositionable {
 
   /**
    * Creates the button used to get the previous block in the list.
+   *
    * @returns The previous button.
    */
   protected createPreviousBtn(): HTMLButtonElement {
@@ -249,6 +255,7 @@ export class WorkspaceSearch implements Blockly.IPositionable {
 
   /**
    * Creates the button used for closing the search bar.
+   *
    * @returns A button for closing the search bar.
    */
   protected createCloseBtn(): HTMLButtonElement {
@@ -257,6 +264,7 @@ export class WorkspaceSearch implements Blockly.IPositionable {
 
   /**
    * Creates a button for the workspace search bar.
+   *
    * @param className The class name for the button.
    * @param text The text to display to the screen reader.
    * @returns The created button.
@@ -271,6 +279,7 @@ export class WorkspaceSearch implements Blockly.IPositionable {
 
   /**
    * Add event listener for clicking and keydown on the given button.
+   *
    * @param btn The button to add the event listener to.
    * @param onClickFn The function to call when the user clicks on
    *      or hits enter on the button.
@@ -296,6 +305,7 @@ export class WorkspaceSearch implements Blockly.IPositionable {
   /**
    * Returns the bounding rectangle of the UI element in pixel units relative to
    * the Blockly injection div.
+   *
    * @returns The component’s bounding box. Null in this
    *     case since we don't need other elements to avoid the workspace search
    *     field.
@@ -308,6 +318,7 @@ export class WorkspaceSearch implements Blockly.IPositionable {
    * Positions the workspace search field.
    * It is positioned in the opposite corner to the corner the
    * categories/toolbox starts at.
+   *
    * @param metrics The workspace metrics.
    * @param savedPositions List of rectangles that
    *     are already on the workspace.
@@ -342,6 +353,7 @@ export class WorkspaceSearch implements Blockly.IPositionable {
 
   /**
    * Handles a key down for the search bar.
+   *
    * @param e The key down event.
    */
   private onKeyDown(e: KeyboardEvent) {
@@ -361,6 +373,7 @@ export class WorkspaceSearch implements Blockly.IPositionable {
 
   /**
    * Opens the search bar when Control F or Command F are used on the workspace.
+   *
    * @param e The key down event.
    */
   private onWorkspaceKeyDown(e: KeyboardEvent) {
@@ -388,6 +401,7 @@ export class WorkspaceSearch implements Blockly.IPositionable {
 
   /**
    * Sets the placeholder text for the search bar text input.
+   *
    * @param placeholderText The placeholder text.
    */
   setSearchPlaceholder(placeholderText: string) {
@@ -399,6 +413,7 @@ export class WorkspaceSearch implements Blockly.IPositionable {
 
   /**
    * Changes the currently "selected" block and adds extra highlight.
+   *
    * @param index Index of block to set as current. Number is wrapped.
    */
   protected setCurrentBlock(index: number) {
@@ -439,6 +454,7 @@ export class WorkspaceSearch implements Blockly.IPositionable {
 
   /**
    * Shows or hides the workspace search bar.
+   *
    * @param show Whether to set the search bar as visible.
    */
   private setVisible(show: boolean) {
@@ -448,6 +464,7 @@ export class WorkspaceSearch implements Blockly.IPositionable {
   /**
    * Searches the workspace for the current search term and highlights matching
    * blocks.
+   *
    * @param searchText The search text.
    * @param preserveCurrent Whether to preserve the current block
    *    if it is included in the new matching blocks.
@@ -472,6 +489,7 @@ export class WorkspaceSearch implements Blockly.IPositionable {
 
   /**
    * Returns pool of blocks to search from.
+   *
    * @param workspace The workspace to get blocks from.
    * @returns The search pool of blocks to use.
    */
@@ -486,6 +504,7 @@ export class WorkspaceSearch implements Blockly.IPositionable {
 
   /**
    * Returns whether the given block matches the search text.
+   *
    * @param block The block to check.
    * @param searchText The search text. Note if the search is case
    *    insensitive, this will be passed already converted to lowercase letters.
@@ -518,6 +537,7 @@ export class WorkspaceSearch implements Blockly.IPositionable {
 
   /**
    * Returns blocks that match the given search text.
+   *
    * @param workspace The workspace to search.
    * @param searchText The search text.
    * @param caseSensitive Whether the search should be case sensitive.
@@ -557,6 +577,7 @@ export class WorkspaceSearch implements Blockly.IPositionable {
   /**
    * Adds "current selection" highlight to the provided block.
    * Highlights the provided block as the "current selection".
+   *
    * @param currentBlock The block to highlight.
    */
   protected highlightCurrentSelection(currentBlock: Blockly.BlockSvg) {
@@ -566,6 +587,7 @@ export class WorkspaceSearch implements Blockly.IPositionable {
 
   /**
    * Removes "current selection" highlight from provided block.
+   *
    * @param currentBlock The block to unhighlight.
    */
   protected unhighlightCurrentSelection(currentBlock: Blockly.BlockSvg) {
@@ -575,6 +597,7 @@ export class WorkspaceSearch implements Blockly.IPositionable {
 
   /**
    * Adds highlight to the provided blocks.
+   *
    * @param blocks The blocks to highlight.
    */
   protected highlightSearchGroup(blocks: Blockly.BlockSvg[]) {
@@ -586,6 +609,7 @@ export class WorkspaceSearch implements Blockly.IPositionable {
 
   /**
    * Removes highlight from the provided blocks.
+   *
    * @param blocks The blocks to unhighlight.
    */
   protected unhighlightSearchGroup(blocks: Blockly.BlockSvg[]) {