+++ plugins/theme-deuteranopia/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.0.3](https://github.com/google/blockly-samples/compare/@blockly/theme-deuteranopia@5.0.2...@blockly/theme-deuteranopia@5.0.3) (2023-11-02)
+
+**Note:** Version bump only for package @blockly/theme-deuteranopia
+
+
+
+
+
+## [5.0.2](https://github.com/google/blockly-samples/compare/@blockly/theme-deuteranopia@5.0.1...@blockly/theme-deuteranopia@5.0.2) (2023-10-30)
+
+**Note:** Version bump only for package @blockly/theme-deuteranopia
+
+
+
+
+
 ## [5.0.1](https://github.com/google/blockly-samples/compare/@blockly/theme-deuteranopia@5.0.0...@blockly/theme-deuteranopia@5.0.1) (2023-08-17)