+++ plugins/toolbox-search/CHANGELOG.md
@@ -3,6 +3,31 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.1](https://github.com/google/blockly-samples/compare/@blockly/toolbox-search@1.2.0...@blockly/toolbox-search@1.2.1) (2023-11-02)
+
+**Note:** Version bump only for package @blockly/toolbox-search
+
+
+
+
+
+## [1.2.0](https://github.com/google/blockly-samples/compare/@blockly/toolbox-search@1.1.6...@blockly/toolbox-search@1.2.0) (2023-10-30)
+
+
+### Features
+
+* add publish types for toolbox-search ([#1950](https://github.com/google/blockly-samples/issues/1950)) ([addae02](https://github.com/google/blockly-samples/commit/addae0292602136d6cb484410ac60e9f7ee3bc3a))
+* Index dropdown field option for toolbox-search ([#1971](https://github.com/google/blockly-samples/issues/1971)) ([036f238](https://github.com/google/blockly-samples/commit/036f23834d74fd108fc4afe3e2032ae4c93cf20d))
+
+
+### Bug Fixes
+
+* **packaging:** Don't advertise non-existent ESM entrypoints ([#2022](https://github.com/google/blockly-samples/issues/2022)) ([d019198](https://github.com/google/blockly-samples/commit/d0191984399b784e2928b8fb4c58257bfa857655)), closes [#1877](https://github.com/google/blockly-samples/issues/1877)
+* update tsconfigs to publish types accurately ([df4b084](https://github.com/google/blockly-samples/commit/df4b0844af712f5025a2ec842458b828f3147676))
+* use includes rather than excludes ([40efc25](https://github.com/google/blockly-samples/commit/40efc255329e3ca476ccc247b95a2d05dd77b45e))
+
+
+
 ## [1.1.6](https://github.com/google/blockly-samples/compare/@blockly/toolbox-search@1.1.5...@blockly/toolbox-search@1.1.6) (2023-09-21)