+++ plugins/typed-variable-modal/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.0.6](https://github.com/google/blockly-samples/compare/@blockly/plugin-typed-variable-modal@7.0.5...@blockly/plugin-typed-variable-modal@7.0.6) (2023-11-02)
+
+**Note:** Version bump only for package @blockly/plugin-typed-variable-modal
+
+
+
+
+
+## [7.0.5](https://github.com/google/blockly-samples/compare/@blockly/plugin-typed-variable-modal@7.0.4...@blockly/plugin-typed-variable-modal@7.0.5) (2023-10-30)
+
+**Note:** Version bump only for package @blockly/plugin-typed-variable-modal
+
+
+
+
+
 ## [7.0.4](https://github.com/google/blockly-samples/compare/@blockly/plugin-typed-variable-modal@7.0.3...@blockly/plugin-typed-variable-modal@7.0.4) (2023-09-14)
 
 **Note:** Version bump only for package @blockly/plugin-typed-variable-modal